refactor: restructure modal components
- Split Modals.vue into individual modal components for better maintainability
- Add new modal components: AssignStudent, ClearLogs, Logout, and StudentInfo
- Remove deprecated Modals.vue component
